split header source
